+++ bc/search/tests/fixtures.py
@@ -1,0 +1,6 @@
+import factory
+
+
+class TermFactory(factory.django.DjangoModelFactory):
+    class Meta:
+        model = "search.Term"