+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## Unreleased
 
 - Add support for Elasticsearch backend
+- Add configurable search synonyms
+- Use Porter stemming algorithm search filter, for English grammar
+- Add an English stop words search filter
 - Add Wagtail admin report of pages missing SEO metadata, and those with unpublished changes
 
 ## 26.0 (2020-09-22)